requirements-dev.txt
anyio==3.3.4
asgiref==3.4.1
attrs==21.2.0
black==21.10b0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
click==8.0.3
coverage==6.2
cryptography==38.0.3
dataclass-type-validator==0.1.1
ecdsa==0.18.0
fastapi==0.70.0
freezegun==1.1.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.2
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
python-jose==3.3.0
regex==2021.11.10
requests==2.26.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
